Added tests to very RAT-362
Added EnvVar and associated processing to consolidate envirnoment variables
Added testing FileUtils
Added OptionCollectionTest.OptionTest functions to create named OptionTest to aid in debugging testing issues.
Added OptionCollecitonTest.TestFunction annotation to support annotations to build the test suite.
