+++ apache-rat-core/src/it/resources/ReportTest/RAT_147/verify.groovy
