+++ apache-rat-core/src/it/resources/ReportTest/RAT_211/verify.groovy
