+++ apache-rat-core/src/it/resources/ReportTest/RAT_362/commandLine.txt
@@ -1,0 +1,5 @@
+--output-style
+xml
+--input-exclude-parsed-scm
+GIT
+--
