+++ apache-rat-core/src/it/resources/ReportTest/RAT_362/expected-message.txt
@@ -1,0 +1,1 @@
+Issues with UNAPPROVED
