+++ apache-rat-core/src/it/resources/ReportTest/RAT_362/src/.gitignore
@@ -1,0 +1,2 @@
+/foo.md
+/src.md
