+++ apache-rat-core/src/it/resources/ReportTest/RAT_81/verify.groovy
